switch to dynamic import
